requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
python-decouple>=3.8
Pillow>=10.0.0
stripe>=6.0.0
whitenoise>=6.5.0
gunicorn>=21.0.0
psycopg2-binary>=2.9.7
django-ckeditor>=6.7.0
django-storages>=1.14.0
boto3>=1.28.0